bug fix: non-integer simValues[slider].Default

+++ js.cs
@@ -54,7 +54,7 @@ namespace blekenbleu
 			 ; 100 (1)		0 - 100
 			 ; 1000 (10)	0 - 1000
 			 */
-			if (100 < Convert.ToInt32(simValues[slider].Default))
+			if (100 < Convert.ToDouble(simValues[slider].Default))
 			{
 				SliderFactor[0] = 10;	// slider to value
 				SliderFactor[1] = 0.1;	// value to slider
